Add copilot toggle button
