+++ src/Explorer/QueryCopilot/Shared/QueryCopilotClient.ts
@@ -21,6 +21,7 @@ import { userContext } from "UserContext";
 import { queryPagesUntilContentPresent } from "Utils/QueryUtils";
 import { useQueryCopilot } from "hooks/useQueryCopilot";
 import { useTabs } from "hooks/useTabs";
+import * as StringUtility from "../../../Shared/StringUtility";
 
 export const SendQueryRequest = async ({
   userPrompt,
@@ -165,7 +166,7 @@ export const OnExecuteQueryClick = async (): Promise<void> => {
 
 export const QueryDocumentsPerPage = async (
   firstItemIndex: number,
-  queryIterator: MinimalQueryIterator,
+  queryIterator: MinimalQueryIterator
 ): Promise<void> => {
   try {
     useQueryCopilot.getState().setIsExecuting(true);
@@ -173,8 +174,7 @@ export const QueryDocumentsPerPage = async (
     useTabs.getState().setIsQueryErrorThrown(false);
     const queryResults: QueryResults = await queryPagesUntilContentPresent(
       firstItemIndex,
-      async (firstItemIndex: number) =>
-        queryDocumentsPage(QueryCopilotSampleContainerId, queryIterator, firstItemIndex),
+      async (firstItemIndex: number) => queryDocumentsPage(QueryCopilotSampleContainerId, queryIterator, firstItemIndex)
     );
 
     useQueryCopilot.getState().setQueryResults(queryResults);
@@ -184,15 +184,20 @@ export const QueryDocumentsPerPage = async (
       correlationId: useQueryCopilot.getState().correlationId,
     });
   } catch (error) {
+    const isCopilotActive = StringUtility.toBoolean(
+      localStorage.getItem(`${userContext.databaseAccount?.id}-queryCopilotToggleStatus`)
+    );
     const errorMessage = getErrorMessage(error);
     traceFailure(Action.ExecuteQueryGeneratedFromQueryCopilot, {
       correlationId: useQueryCopilot.getState().correlationId,
       errorMessage: errorMessage,
     });
-    useQueryCopilot.getState().setErrorMessage(errorMessage);
     handleError(errorMessage, "executeQueryCopilotTab");
     useTabs.getState().setIsQueryErrorThrown(true);
-    useQueryCopilot.getState().setShowErrorMessageBar(true);
+    if (isCopilotActive) {
+      useQueryCopilot.getState().setErrorMessage(errorMessage);
+      useQueryCopilot.getState().setShowErrorMessageBar(true);
+    }
   } finally {
     useQueryCopilot.getState().setIsExecuting(false);
     useTabs.getState().setIsTabExecuting(false);
